Engine: Group 'bad' case handling for optional chains
This patch changes the way optional chains are dealt with in the
bytecode. Instead of dealing with the 'bad' case (where the base of the
lookup is null or undefined) of each instruction separately, all
optional operations point to the same piece of code at the end of the
optional chain to deal with bad accesses.

In practice, for the lookup `root?.foo.bar?.baz` the following
bytecode instructions are generated.

    LoadQmlContextPropertyLookup  // root
    GetOptionalLookup --v         // ?.foo
    GetLookup           |         //  .bar
    GetOptionalLookup --v         // ?.baz
    Jump done ------------v
undefined:         <----< |
    LoadUndefined         |
done:              <------<

In this way, the 'bad' case is handled in one place at the undefined
label. If, on the other hand, the chain evaluation reaches the bottom,
one jump takes the resulting value to the rest of the program. In this
way, the 'bad' case has a constant size relative to the length of the
chain. If no optional operation is performed at all. The 'bad' case
handler is not generated at all.

For this to work, GetOptionalLookup now jumps to the undefined label
when its base is null of undefined. Other operations such as function
calls `f?.()`, array access `a?.[0]` and delete expressions
`delete foo?.bar` have also been adapted to point to the undefined
label.

Change-Id: I07158efc8767d84a7588299cae9fb763b0f6e253
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>