ListView: position delegate items at zero on the inactive axis
FxListItemSG::setPosition() only cares about a single axis, which is
based on the current orientation. If the orientation is Vertical, then
the y axis will be used with positioning the delegate item, and if the
orientation is Horizontal, then the x axis will be used instead.

This would cause problems when the ListView is changing its orientation.
Because if the ListView's orientation suddently changes from Vertical to
Horizontal, then all the delegate items will keep their old y value,
which FxListItemSG::setPosition() won't do anything about since the new
"active" axis is now the x axis.

This patch fixes the issue by resetting the value for the "inactive"
axis back to zero, in pointForPosition() which is used by
FxListItemSG::setPosition().

The problem wasn't noticeable when using a normal QQmlDelegateModel,
since the delegate items would be created lazily when they needed to
appear inside the viewport, at which point they'd be given the correct x
and y values. (The position is first (0, 0), and then layoutVisibleItems
will updated one of those values based on the orientation).
But if the user decides to use a QQmlObjectModel based model instead,
then the delegate items would all be created immediately with a
persistent lifetime, which would never have their position reset as a
natural result of the item view delegate lifecycle mechanism.

Fixes: QTBUG-115696
Pick-to: 6.5
Change-Id: I5377aeb556be2f536794f489b8232c5b9675ab78
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit f03a9839b689a3a1810846bb8ec08e02a9bf23b5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>