Drawer: adjust opening edge to the rotation of the window's content item
If the content of a window is rotated, then the drawer needs to adjust
the edge that it opens from. We can only meaningfully handle rotations
by multiple of 90 degrees.

[ChangeLog][Qt Quick][Drawer] If the window's content item is rotated, then
the drawer will adapt the edge from which it can be drawn out accordingly.
Only multiple of 90 degrees are supported.

Fixes: QTBUG-80910
Task-number: QTBUG-71117
Task-number: QTBUG-115536
Pick-to: 6.6
Change-Id: I213b988053e170d198c2f180de85107edf83c812
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>