StyleGenerator: allow the user to select which controls to generate
Typically a designer will not produce a style for every singel
control in Figma. But the generator still generates QML and images
for them all regardless. This gives a long generate/download time, and
produce a style that will use the "default" figma design for the
remaining controls, instead of using controls from the fallback style.

Therefore, based on the controls listed in the config.json file, present
a list to the user that lets him choose what to generate.

Showing a list of controls to the user in the UI takes quite some
space. Therefore, take the opportunity to refactor the UI a bit
by dividing it into more QML files, and use a ScrollView to
combine the invidual pieces.

Change-Id: Ib8af94c53a3681b7ff57ab8a6b8eef2e87fba471
Reviewed-by: Doris Verria <doris.verria@qt.io>