DelegateModel: don't insert items using an invalidated iterator
The code in the while-loop in insert() has to allocate memory, and that
might trigger the garbage collector, which will clear the cache. This
invalidates the "before" iterator passed into insert(), resulting in
asserts when inserting new items into the cache.

To fix this we have to detect that the `before` iterator became invalid,
and then recompute the iterator. Change the return value of the insert()
function to provide more information than true/false, and use the logic
and information at the call-site to retry the insertion.

As a drive-by, skip computations of local values if they are not going
to be used.

Fixes: QTBUG-119181
Pick-to: 6.5
Change-Id: I71543b8df776dbeaed5a275960ad34df1d12e19f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 4c1cee7e116012ed34ec1e76331736b89f1bb374)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>