StyleGenerator: remove unused QML API from the bridge
The 'overwriteQml' and 'verbose' options are no longer
used from QML. So remove the API from the bridge
to clean up a bit.

In case of overwriteQml, there is no need to offer this,
since the QML files will not (from StyleGenerator's
POW) change. The developer should instead remove the QML files
manually from the style folder (or remove the whole folder) in
order to regenerate them.

Change-Id: I6dea2e84db4e962e462636b398bcf4cb35e9cf82
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>