Drag'n'drop: fix support for string data that is not a "text/" mime type
Amends 17318c4805d089f10c25f8165dae7d1f908bd311, after which mime data
that was provided as a QString, but not of mime type "text/", was
ignored with a warning. This breaks use cases where mime type of e.g.
"application/json" type, which is then provided as a string rather than
a byte array.

This is then behavior-compatible with pre-6.5 releases. We still warn if
the mime type is "text/" with an encoding we don't know about. Remove
the test data set that now does produce a valid mime data (image/png mime
with data provided as string), and add a test case for application/json
instead.

Fixes: QTBUG-119216
Pick-to: 6.6 6.5
Change-Id: I1c321cd92f45ea479ee9abc838a44828d1300433
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>