Use Qt::WindowNoState for child windows
The Qt::WindowState enum is documented to specify the current state of
top-level windows, and do not make sense for child windows. It's up to
the client code to ensure any sizing of child windows/widgets, e.g. via
layouts.

Many of our QPA backends bail out on QPlatformWindow::setWindowState()
for a child window. Ideally we would do this at a higher level, in
QWindow, but that's a bigger task, and the semantics of what happens
when a window is moved from being top level to child or back are not
fully clear.

As a first step, we ensure that the default window state when showing
a child window is Qt::WindowNoState.

Unfortunately, the QPlatformIntegration::defaultWindowState() API only
takes the window flags into account, which sadly do not have a way to
reflect whether the window is a child window or not (Qt::SubWindow is
not a child window, see QTBUG-115729).

We don't want to pass a QWindow to this API, as it would mean QWidget
would need to create a window when requesting the default window state.

Instead we hard-code the opt-out for child windows.

Change-Id: I2bb9b0f3b9f42351e065a7802d647b613f624993
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>