Long live incremental garbage collection in QML!
The design of the garbage collector is described in
src/qml/memory/design.md.

The gc and gcdone test helpers are adjusted to drive the gc to
completion, even when in incremental mode.

Parts of tst_qv4mm and tst_qqmlqt need to run with the incremental gc
disabled, as they call gc inside QML and assumes that the GC finishes
before returning.

Initial-patch-by: Rafal Chomentowski <rafal.chomentowski@ge.com>
Task-number: QTBUG-119274
Change-Id: I1d94f41bc7a434fad67de0fd46454b6db285f2eb
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>