Accept mouse event if grabbed after filtering in PathView
After PathView is flicked, it can be stopped by a mouse press.
Normally, it filters mouse events to control the
movement within the view. The press after flick is grabbed and
consumed to stop movement; but it didn't accept this event, so
delivery was continuing to other items beneath. Now we accept
the mouse press so that QQuickDeliveryAgent stops propagating it.

Fixes: QTBUG-115121
Pick-to: 6.5
Change-Id: Id4119b94c004883ff40d728f78b1af729290f331
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit d739b57098e8c15a04f41658e5a19f2eadebef44)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>