Fix text decoration on software backend
This amends d0cb311e0d3cd84c933bd993e6003f45be2ebdd8. In said
change, text decoration was implemented using the curve renderer
shaders, but neglected to consider that the QSGInternalTextNode is
used on all backends. This caused decorations to disappear when
using the software backend.

This patch slightly refactors it, so that the shader-based approach
is only used on the default RHI-based backend, and the software
backend will still use a rectangle node, since this works well with
that backend, even on fractional scaling displays.

Task-number: QTBUG-86977
Change-Id: I08a595b1a794c5b41a0f239328dc6f5f2fce70c0
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>