StyleGenerator: update code to generate icons as individual items
Using a single Component in Figma for all the icons turned out
to be a mistake. Designers like to "swap" out the icons in a
control instance, rather than choosing it from a list of
component variations. This also opens up for using icons from
outside of the templates, and e.g from an external library.

The Icons dialog in Figma has therefore been updated to be just a
container where you can drop individial icons as needed (and
perhaps none at all). This patch will update the code in
StyleGenerator to handle this change. The name of the icons
will then just be the name of the icon in Figma, and no more
trying to resolve the name based on variation properties.

Change-Id: I65ea822141b4a95aace0d2910086c682185eaaae
Reviewed-by: Doris Verria <doris.verria@qt.io>