qmlls: completions in switch/case/default statements
Add completions for switch/case/default statements, add the needed
locations and use those locations to provide the autocompletion

Also fix the suggestion of case and default statements to not happen
inside of statements nested inside a case or default statement.
Instead, only suggest them inside the {} of a switch statement and
at the end of a case or default statement.

Add some tests in a separate file.

Task-number: QTBUG-117445
Change-Id: Ib2c8e7a88b20d5c6360fa8a491585265f42c4146
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>