CMake: Fix <module>_qmllint for modules without qml files on Windows
If have_qmllint_files is empty, pass a dummy COMMAND to
add_custom_target(), instead of just an empty string. While an
empty string does seem to work on Linux and macOS, it will lead to
an error on Windows:

   CreateProcess failed: The system cannot find the file specified.

This was at least the case with CMake version 3.26.0, and ninja
version 1.11.1.

Pick-to: 6.5 6.6
Fixes: QTBUG-115710
Change-Id: Ifb3572d695b836e603572642c8aac8bf5bb8214d
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>