Fix IconImage test on systems with mixed DPR displays
When running on a system with two displays with different DPR values, it
used the highest value of all systems (which is what
QGuiApplication::devicePixelRatio returns). Some tests then failed when
the test's windows opened on the lower-DPR display, and other tests
failed when the windows opened on the high-DPR display.

Fix this by initializing the DPR values in each test function after the
view is shown, reading the DPR value from that view. Fold the skipping
if the DPR is too high into that macro.

Pick-to: 6.6
Change-Id: I2b2f3540e26ee0c65d7ef62dd9b11640b5ca0274
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>