Fix possible crash when using CurveRenderer text
The curve renderer would keep references to its cache, which
could be invalid if the font database had been purged.

For the distance field renderer we handle this by never purging
caches that are in use, but we don't really need such a mechanism
with the CurveRenderer, since we don't need to keep any data
around for the actual rendering later (like the textures in the
other caches).

Instead, we just look up the cache when we need it, if it has
been purged in the mean time, it's just re-created.

Change-Id: I0c2a84f3176787f94c599b65296b9d87d2e28858
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>