Fix text decorations with software backend on fractional scale
We were drawing the text decorations twice, once through the
cross-backend mechanism in the text node and then on top with QPainter.
However, when drawing it through QPainter the fractional scale was
not handled correctly.

There may be a bug in QPainter::drawGlyphRun(), but for the software
backend in particular we should just rely on the cross-backend
code for this, so we disable the decoration on the QGlyphRun as a
quick fix.

Fixes: QTBUG-99266
Change-Id: Ie19d3554fc03dae6a6cbf219e91d5bb46fb3f15c
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Vladimir Belyavsky <belyavskyv@gmail.com>