QQuickItem: Call markSortedChildrenDirty before QQuickItemPrivate::dirty()
Marking an item dirty will normally result in adding the item to the
dirty list, and telling the window about it, via QQuickWindowPrivate's
dirtyItem(), which in turn does QQuickWindow::maybeUpdate().

This results in an asynchronous update, where we then render the scene
graph, including updating its paint nodes.

But QQuickItemPrivate::dirty() may also involve synchronous actions,
such as calling QQuickItemPrivate::transformChanged().

To be on the safe side, we markSortedChildrenDirty before dirtying,
so that any synchronous logic that runs as part of the dirtying will
observe the newly dirtied sorted children.

We were already doing this for QQuickItemPrivate::ChildrenChanged,
but not for QQuickItemPrivate::ChildrenStackingChanged.

Change-Id: I47f0ee81e17ed766ead7c7644596e90d240482a2
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>