QQuickFocusFrame: don't try to create frame if context's engine is null
In the case of QQuickWebEngineView and its tests, the item gets focus
even though it has no QQmlEngine associated with its context. We need
the engine for creating the focus frame component, so return early if
it's null.

There's no test for this because it's not easily reproduced outside of
qtwebengine.

Fixes: QTBUG-118750
Pick-to: 6.5 6.6 6.7
Change-Id: I3b96b1c83944f31a6ed80b3c95256bb5c6e6b3bf
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Anu Aliyas <anu.aliyas@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>