qmlls: find usages of enumerations
Enumerations in Qml are bound to file component they are defined in. Qml
engine doesn't allow to have same named enum in the same scope, but it
allows having the same named enum within another scope. Nevertheless,
they have no way to be accessed, for example it is not possible to use
<FileName>.<EnumValue> for an enum declared inner scope.

Consider this limitation, don't find the enum values defined in inner
scopes.

Introduce finding enum names. Also fix a completion test that was using
enums with incorrect semantics. This has to be squashed into this commit
otherwise test would break.

Change-Id: I1c4f46dd3ee1de4e04021cb35614a15be566105d
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>