Ensure ItemObservesViewport works when item is moved to new ancestor
When an item has ItemObservesViewport set, it will receive
transformChanged callbacks when any of its ancestors changes their
position, size, etc.

This relies on every ancestor having the subtreeTransformChangedEnabled
flag set, which was previously ensured when the item called setFlag().

But when the item then moved to a different ancestor hierarchy, we
would only set subtreeTransformChangedEnabled on the new parent,
but not any of its ancestors, which meant that the item would
no longer pick up changes beyond the immediate parent.

We would also not send a transformChanged in response to the reparenting
itself, which is critical, as the item's clip or scene positions may
have changed in response to the new parent.

Both of these cases are now handled as part of the ItemChildAddedChange
of the new parent.

Change-Id: I5000fb46d7638ce9674d4879890cf4344e5ce538
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>