JSEngine: Optimize QV4InternalClass for tests on MSVC
Running tst_ecmascripttests on MSVC in debug mode has become very slow
and times out the CI. Analysing the program reveals that most of the
slowdown is due to extensive wait times caused by locks. The test is run
on all threads by default and this constant synchronization tanks
performance.

A large amount of these locks come from the fact that, in debug mode,
MSVC standard containers use locks for most operations on iterators.

This patch changes the container of the internal class transitions from
an std::vector to a QVarLengthArray. This eliminates the iterator lock
problem. The QVarLengthArray is given one inline entry to store a value.
From the tests, it seems that the transitions contain only 0 or 1
entries about 84% of the time. In the remaining cases, more allocations
will be needed. The additional entry will also increase the size of the
containing object by 24 bytes. This should be a worthwhile tradeoff.
This change alone has a significant impact on the duration of the tests.

tst_ecmascripttests on 13900k with 32 threads
                 Debug MSVC Windows      Debug GCC Linux
baseline:            2267s                  104s
QVarLengthArray       569s (~ -73%)         102s (~ -2%)

This should be enough to no longer timeout the CI but many issues still
remain.

Change-Id: I69fefabc0375d76c817ef7d2d3b2e97dc1ace5bc
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>