qmlls: fix method body and statement completion
It seems that the Dom construction was not creating a block for empty
function bodies. This led to a bug where completion was not working inside
an empty method bodies, so fix it in this commit. Add a test in
tst_qmldomitem for this case.

Also, it seems that the (non-empty) function bodies had the wrong
SourceLocations: instead of starting and ending at the `{}` tokens, they
were starting at the start and end of the statementlist.
This was making the completion stop working, for example in
```
function f() {        return 3;        }
                      <-body-->  (old source location)
             <----------body----------->  (new source location)
             <------->        <--------> (where completion failed)
```
you would not get any JS statement completion in the whitespace
before or after the return statement.

Also, use the CompletionContext when completing JS statements. This is
because expression statements need the context to provide correct
autocompletion of field member expressions like
`root.<completion here>`.

Add a test for both cases, and replace the QEXPECT_FAIL test
"inMethodBody" with "insideMethodBody". There is no need to test for
keywords like "for", "while", and co anymore: these JS statements
are already being proposed as snippets, and these snippets already have
their own tests.

Task-number: QTBUG-117445
Change-Id: Ia9157087524fbb1fc0f9cf64502887b4630af6b9
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>