Improve hash strategy for font caches
We were using a string key for font lookups instead of
creating a proper type for this. This was a quick fix
which has grown over time and is inconvenient and slow.
Instead we now introduce a FontKey type which uses
Qt 6's qHashMulti() for this, also making it
automatically adapt to changes in the FaceId as well.

Change-Id: I32dfe3b809710e3d76896a5ddef03649bb4da0d5
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>