Introduce QQuickView loadFromModule
...as well as its ctor equivalent. This unifies the API with what is
provided in QQmlApplicationEngine and QQmlComponent.

If the type is composite, we update source accordingly. C++ defined
types don't have any source, so we leave source empty – this is fine as
source only promises to return something if it has been set.
Typename and URI can't be retrieved after they have been passed to
QQuickView. This could be changed later if a use case for it arrives.

Fixes: QTBUG-111727
Change-Id: I9f8a27a6bf9132a37a2d3d09f90fbea7803342ee
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>