StyleGenerator: clean-up CLI options
We should only parse CLI options when running in CLI
mode. We can therefore move the code block that deals
with that into the block that handles CLI.

At the same time, clean up the naming of the some of
the CLI options, and give them sensible defaults when
left unspesified from the command line. The --generate
will now also follow the same API as from QML, where you
specify individual controls rather than a regexp.

Change-Id: I6ee862c29e83b18b27a94beff91db756d9028d89
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>