Android: Add signal handling feature to QtQuickView
Add add/remove signal handler functions to QtQuickView
These signal handlers can cover parameterless signals
as well as signals with basic types. If the notifier signal
of a property is parameterless, the internal handlers try
to read and convert the property value to the Java type and
pass it to the lambda handler.

Pick-to: 6.7
Task-number: QTBUG-119751
Change-Id: I249ab102b84e92d1595ebfb0a8822bda1f4ebc23
Reviewed-by: Rami Potinkara <rami.potinkara@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>