qmlls: completions for comma expressions
Represent comma expressions as BinaryExpressions in the Dom, as comma
expressions behave very similarly to binary expressions. This means that
completions for comma expressions can reuse the already existing
completion for binary expression, and no new code has to be added for
that.
Still, add a test to make sure completion works as expected, and another
to make sure the dom representation is correctly constructed.

Task-number: QTBUG-92876
Task-number: QTBUG-119798
Change-Id: I7c61abaa4dc6a153e318322e4d64211231e63e3a
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>