Allow QQuickRt adopt a native texture array for multiview rendering
Limiting multiview to fromRhiRenderTarget() is quite limiting, since
in practice it often involves duplicating the convenience logic that
is built into Qt Quick for fromVulkanImage() and similar.

We should also support passing in a texture array object to from
fromOpenGLTexture, fromVulkanImage, etc. The array size then
implies the view count.

We are also required to use a texture array for depth/stencil as
well in this case.

Add the necessary plumbing so that clients, such as an OpenXR
integration, can continue to use fromXxxxx(), just the
*MultiView variants, instead of forcing them to create a
QRhiTextureRenderTarget themselves upfront.

Task-number: QTBUG-119850
Task-number: QTBUG-114871
Change-Id: I79978e5ff5f154c8914afbe2478d002291219aaf
Reviewed-by: Andy Nichols <andy.nichols@qt.io>