CMake: Fix the all_qmllint* targets for VS generators, take II
When using the VS generator, the foo_qmllint targets were run on every
build, which is neither expected nor wanted. The reason for this is a
pecularity in MSBuild and as a reaction to this, special behavior of
CMake: add_dependencies(A B) will enable B in the default build of the
solution - even if B is not triggered by the ALL target. See upstream
CMake issue #16668 for details.

We now work around this problem by building foo_qmllint with "cmake
--build" in the COMMANDs of all_qmllint. As a consequence, we must
create all_qmllint, all_qmllint_json, and all_qmllint_module at the end
of the project directory scope.

Since we now require cmake_language(DEFER CALL), users who generate VS
projects with CMake versions older than 3.19 will see a warning that the
all_qmllint* targets cannot be created. The VS installation comes with a
recent enough CMake version (on my machine it's 3.26.4), so this should
not be a problem.

[ChangeLog][CMake] When using CMake's Visual Studio project generator,
the creation of the targets all_qmllint, all_qmllint_json and
all_qmllint_module requires now CMake 3.19 or newer. A warning is
printed for older CMake versions. This warning can be disabled by
setting QT_NO_QMLLINT_CREATION_WARNING.

Pick-to: 6.5 6.6 6.7
Fixes: QTBUG-115166
Task-number: QTBUG-115166
Change-Id: I0bd704caa3f36cfda739425c2126a0ec7ac4f85e
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>